refactor: update article category
