+++ src/actions/cms/articleCategories/update.ts
@@ -1,12 +1,12 @@
 'use server'
 import { articleCategorySchema } from './schema'
-import prisma from '@/prisma'
-import { createPrismaActionError, createZodActionError } from '@/actions/error'
+import { createZodActionError } from '@/actions/error'
 import type { ArticleCategorySchemaType } from './schema'
 import type { ActionReturn } from '@/actions/Types'
 import type { ExpandedArticleCategory } from '@/cms/articleCategories/Types'
+import { updateArticleCategory } from '@/server/cms/articleCategories/update'
 
-export async function updateArticleCategoryVisibility(
+export async function updateArticleCategoryVisibilityAction(
     // disable eslint rule temporarily until function is implemented
     // eslint-disable-next-line @typescript-eslint/no-unused-vars
     id: number,
@@ -16,33 +16,12 @@ export async function updateArticleCategoryVisibility(
     throw new Error('Not implemented')
 }
 
-export async function updateArticleCategory(
+export async function updateArticleCategoryAction(
     id: number,
     rawData: FormData | ArticleCategorySchemaType
 ): Promise<ActionReturn<ExpandedArticleCategory>> {
     const parse = articleCategorySchema.safeParse(rawData)
-
-    if (!parse.success) {
-        return createZodActionError(parse)
-    }
-
-    const { name, description } = parse.data
-
-    try {
-        const articleCategory = await prisma.articleCategory.update({
-            where: {
-                id
-            },
-            data: {
-                name,
-                description
-            },
-            include: {
-                articles: true
-            }
-        })
-        return { success: true, data: articleCategory }
-    } catch (error) {
-        return createPrismaActionError(error)
-    }
+    if (!parse.success) return createZodActionError(parse)
+    const data = parse.data
+    return await updateArticleCategory(id, data)
 }
